% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airports.R
\name{osn_fetch_around_airport}
\alias{osn_fetch_around_airport}
\title{Fetch state vectors around an airport}
\usage{
osn_fetch_around_airport(ident, radius_nm, date = Sys.Date() - 1, con = NULL)
}
\arguments{
\item{ident}{ICAO airport code (e.g. \code{"EHAM"}).}

\item{radius_nm}{Radius in Nautical Miles.}

\item{date}{Date to fetch. Default: yesterday.}

\item{con}{A DBI connection from \code{\link[=osn_connect]{osn_connect()}}. If \code{NULL}, a new
connection is created automatically.}
}
\value{
Filtered lazy \code{tbl_dbi}.
}
\description{
Convenience wrapper that combines \code{\link[=osn_fetch_day]{osn_fetch_day()}},
\code{\link[=osn_airport_coords]{osn_airport_coords()}}, and \code{\link[=osn_filter_radius]{osn_filter_radius()}}.
}
