% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{osn_disconnect}
\alias{osn_disconnect}
\title{Disconnect from OpenSky Network}
\usage{
osn_disconnect(con)
}
\arguments{
\item{con}{A DBI connection returned by \code{\link[=osn_connect]{osn_connect()}}.}
}
\description{
Shuts down the DuckDB connection cleanly.
}
